+++ pysiral/l2proc.py
@@ -24,6 +24,7 @@ from pysiral.path import filename_from_path
 
 from collections import deque, OrderedDict
 from datetime import datetime
+import numpy as np
 import time
 import sys
 import os
@@ -51,6 +52,7 @@ class Level2Processor(DefaultLoggingClass):
 
         # This variable will contain a list with the auxiliary data handlers
         self._registered_auxdata_handlers = []
+        self._auxhandlers = {}
 
         # Output_handler (can be one or many)
         self._output_handler = product_def.output_handler
@@ -94,13 +96,15 @@ class Level2Processor(DefaultLoggingClass):
 
     @property
     def l2_auxdata_source_dict(self):
-        """ A dictionary that contains the descriptions of the auxiliary
-        data sources """
+        """ A dictionary that contains the descriptions of the auxiliary data sources """
         auxdata_dict = {}
-        for auxdata_type in self.registered_auxdata_handlers:
+        for auxdata_id, auxdata_type in self.registered_auxdata_handlers:
             try:
-                handler = getattr(self, auxdata_type)
-                auxdata_dict[auxdata_type] = handler.longname
+                handler = self._auxhandlers[auxdata_id]
+                if auxdata_type not in auxdata_dict:
+                    auxdata_dict[auxdata_type] = handler.longname
+                else:
+                    auxdata_dict[auxdata_type] = auxdata_dict[auxdata_type]+", "+handler.longname
             except AttributeError:
                 auxdata_dict[auxdata_type] = "unspecified"
         return auxdata_dict
@@ -181,6 +185,7 @@ class Level2Processor(DefaultLoggingClass):
             auxdata_type  = auxdata_dict.keys()[0]
             auxdata_def = auxdata_dict[auxdata_type]
 
+
             # Get options from l2 processor definition file
             # NOTE: These will intentionally override the default options defined in auxdata_def.yaml
             l2_procdef_opt = auxdata_def.get("options", None)
@@ -188,12 +193,15 @@ class Level2Processor(DefaultLoggingClass):
             # Retrieve the class (errors will be caught in method)
             auxhandler = self._auxclass_handler.get_pyclass(auxdata_type, auxdata_def["name"], l2_procdef_opt)
 
+            # Get a unique id of the auxhandler
+            auxhandler_id = "%s_%s" % (auxdata_type, auxhandler.pyclass.lower())
+
             # Add & register the class to the Level-2 processor instance
-            setattr(self, auxdata_type, auxhandler)
-            self._registered_auxdata_handlers.append(auxdata_type)
+            self._auxhandlers[auxhandler_id] = auxhandler
+            self._registered_auxdata_handlers.append((auxhandler_id, auxdata_type))
 
             # Report the auxdata class
-            self.log.info("Processor Settings - %s auxdata handler: %s" % (auxdata_type.upper(), auxhandler.pyclass))
+            self.log.info("Processor Settings - %s auxdata handler registered" % auxhandler_id.upper())
 
     def _report_output_location(self):
         for output_handler in self._output_handler:
@@ -246,6 +254,11 @@ class Level2Processor(DefaultLoggingClass):
                 continue
             l2 = Level2Data(l1b.info, l1b.time_orbit, period=period)
 
+            # Transfer l1p parameter to the l2 data object (if applicable)
+            # NOTE: This is only necessary, if parameters from the l1p files (classifiers) should
+            #       be present in the l2i product
+            self._transfer_l1p_vars(l1b, l2)
+
             # Get auxiliary data from all registered auxdata handlers
             error_status, error_codes = self._get_auxiliary_data(l2)
             if True in error_status:
@@ -254,7 +267,6 @@ class Level2Processor(DefaultLoggingClass):
 
             # Surface type classification (ocean, ice, lead, ...)
             # (ice type classification comes later)
-            # TODO: Add L2 classifiers (ice concentration, ice type)
             self._classify_surface_types(l1b, l2)
 
             # Validate surface type classification
@@ -334,6 +346,32 @@ class Level2Processor(DefaultLoggingClass):
             l1bfilter.set_options(**filter_def.options)
             l1bfilter.apply_filter(l1b)
 
+    def _transfer_l1p_vars(self, l1b, l2):
+        """ Transfer variables from l1p to l2 object"""
+
+        # Make this a backward compatible feature (should work without tag in l2 processor definition file)
+        if not "transfer_from_l1p" in self.l2def:
+            return
+
+        # Get and loop over data groups
+        data_groups, vardefs = td_branches(self.l2def.transfer_from_l1p)
+        for data_group, varlist in zip(data_groups, vardefs):
+
+            # Get and loop over variables per data group
+            var_names, vardefs = td_branches(varlist)
+            for var_name, vardef in zip(var_names, vardefs):
+
+                # Get variable via standard getter method
+                # NOTE: Will return None if not found -> create an empty array
+                var = l1b.get_parameter_by_name(data_group, var_name)
+                if var is None:
+                    var = np.full((l2.n_records), np.nan)
+
+                # Add variable to l2 object as auxiliary variable
+                l2.set_auxiliary_parameter(vardef.aux_id, vardef.aux_name, var, None)
+
+                self.log.info("- Transfered l1p variable: %s.%s" % (data_group, var_name))
+
     def _get_auxiliary_data(self, l2):
         """ Transfer along-track data from all registered auxdata handler to the l2 data object """
 
@@ -345,10 +383,10 @@ class Level2Processor(DefaultLoggingClass):
         auxdata_error_status = []
         auxdata_error_codes = []
 
-        for auxdata_type in self.registered_auxdata_handlers:
+        for (auxdata_id, auxdata_type) in self.registered_auxdata_handlers:
 
             # Get the class
-            auxclass = getattr(self, auxdata_type)
+            auxclass = self._auxhandlers[auxdata_id]
 
             # Transfer variables (or at least attempt to)
             auxclass.add_variables_to_l2(l2)